requirements.txt
Django==5.1.7
asgiref==3.8.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-filter==25.1
PyJWT==2.9.0
sqlparse==0.5.3